requirement.txt
streamlit==1.10.0
python-dotenv==0.19.0
Pillow==8.2.0
google-generativeai==0.0.2
